+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Release 0.9.1
 - Fixed an UDP issue that could imply losing packet if they were sent consequently very fast.
 - Fixed a TCP encoding issue.
+- Fixed a Websocket issue that could imply to lost a message just after being accepted.
 
 ## Release 0.9.0
 - Removed serialization. **Rationale:**